Moved window related keywords back to BrowserManagementKeywords to oversee the relevant changes better.
